hiding away life and chess until a later date
