+++ components/Description.js
@@ -4,10 +4,9 @@ export default function Description(){
     return ( 
         <div className={styles.about}>
             <p> Hello there! My name is Daniel and its lovely to meet you! </p>
-            <p> On this website, its going to be a omnishambles (to put it politely) of different things i try out! Please forgive the trainwreck </p>
-            <p> i am currently working on a number of different projects, including this website, as i learn react and next, along with GCP and a host of other technologies</p>
-            <p> My bread and butter really is python, where at the Uni of Southampton I specialised in machine learning - my projects will be posted under projects, or you can find my cv at the bottom of this page :) </p>
-            Thank you for joining me on this journey through the wild west of technology
+            <p> Please feel free to browse this website as I add new elements (chess board incoming!) and try new things!</p>
+            <p> I am currently working on a number of different projects, including this website, as I learn react and next, along with GCP and a host of other technologies</p>
+            <p> My bread and butter is python, where at the Uni of Southampton I specialised in machine learning - my projects will be posted under projects, or you can find my cv at the bottom of this page :) </p>
         </div>
     )
 }
