+++ components/Footer.js
@@ -1,7 +1,7 @@
 const Footer = () => {
 
     return (
-        <div className="text-center bottom-0 w-full text-white p-6 bg-black">
+        <div className="fixed left-0 bottom-0 right-0 text-center w-full text-white p-6 bg-black">
             Made with 🧡 - 2022 
         </div>
     );
